pyproject.toml
[project]
name = "banking-a2a-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.25",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-instrumentation-fastapi>=0.61b0",
    "opentelemetry-instrumentation-httpx>=0.61b0",
    "opentelemetry-sdk>=1.40.0",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "uvicorn>=0.42.0",
]
